feat: create hook for healthpoints modal
